added a lab7.c file

+++ labs7/lab7.c
@@ -1,12 +1,12 @@
 #include<stdio.h>
 #include<stdlib.h>
 int main(int argc,char*argv[]){
-    int a[100];
+    double a[100];
     int i =1;
     int j = 0;
     int x=0;
     int y = 0;
-    int temp;
+    double temp;
     for(i=1;i<argc;i++,j++){
         a[j]=atof(argv[i]);
     }
@@ -17,6 +17,6 @@ temp=a[x];
 a[x]=a[x+1];
 a[x+1]=temp;}
 }
-printf("The min is：%d\n",a[argc-2]);
+printf("The min is：%g\n",a[argc-2]);
 return 0;
 }
